+++ packages/cfd/src/Containers/cfd-password-modal/cfd-create-password-form.tsx
@@ -1,0 +1,65 @@
+import React from 'react';
+import { FormikHelpers } from 'formik';
+
+import { MultiStep } from '@deriv/components';
+
+import { CFD_PLATFORMS } from '../../Helpers/cfd-config';
+import { TCFDPasswordFormValues } from '../cfd-password-change';
+
+import ChangePasswordConfirmation from '../cfd-change-password-confirmation';
+import CreatePassword from './create-password';
+
+import '../../sass/cfd.scss';
+import { TCFDCreatePasswordFormProps, TMultiStepRefProps, TOnSubmitPassword } from './cfd-password-modal.types.js';
+
+const CFDCreatePasswordForm = ({
+    has_mt5_account,
+    platform,
+    error_message,
+    validatePassword,
+    submitPassword,
+    is_real_financial_stp,
+}: TCFDCreatePasswordFormProps) => {
+    const multi_step_ref = React.useRef<TMultiStepRefProps>();
+    const [password, setPassword] = React.useState('');
+
+    const onSubmit: TOnSubmitPassword = (values, actions) => {
+        if (platform === CFD_PLATFORMS.MT5 && has_mt5_account) {
+            setPassword(values.password);
+            multi_step_ref.current?.goNextStep();
+        } else {
+            submitPassword(values, actions);
+        }
+    };
+
+    const steps = [
+        {
+            component: (
+                <CreatePassword
+                    password={password}
+                    platform={platform}
+                    error_message={error_message}
+                    validatePassword={validatePassword}
+                    onSubmit={onSubmit}
+                    is_real_financial_stp={is_real_financial_stp}
+                />
+            ),
+        },
+        {
+            component: (
+                <ChangePasswordConfirmation
+                    className='cfd-password-modal__change-password-confirmation'
+                    platform={platform}
+                    onConfirm={(_values: TCFDPasswordFormValues, actions: FormikHelpers<TCFDPasswordFormValues>) =>
+                        submitPassword({ password }, actions)
+                    }
+                    onCancel={() => multi_step_ref.current?.goPrevStep()}
+                />
+            ),
+        },
+    ];
+
+    return <MultiStep ref={multi_step_ref} steps={steps} />;
+};
+
+export default CFDCreatePasswordForm;
