+++ packages/cfd/src/types/cfd-store.types.ts
@@ -1,8 +1,8 @@
 import { FormikHelpers } from 'formik';
 import { DetailsOfEachMT5Loginid, Mt5NewAccount, VerifyEmailResponse } from '@deriv/api-types';
 import { TDetailsOfEachMT5Loginid, TTradingPlatformAvailableAccount } from 'Components/props.types';
-import { TCFDPasswordFormValues } from 'Containers/cfd-password-modal';
 import { TDxCompanies, TMtCompanies } from 'Stores/Modules/CFD/Helpers/cfd-config';
+import { TCFDPasswordFormValues } from 'Containers/cfd-password-modal/cfd-password-modal.types';
 
 export type TCFDStore = {
     setMT5TradeAccount: <T>(arg: T) => void;
